requirements.txt
PyQt5
pillow
requests
opencv-python
bcolz
matplotlib
numpy
pandas
scikit-learn
tensorboardX
tqdm

fastapi
uvicorn
sqlalchemy
aiofiles

pyjwt
passlib[bcrypt]